refactor: 重构agent类型和数据结构
